+++ addon/-private/constraints/present.js
@@ -2,11 +2,11 @@ import { isPresent } from '@ember/utils';
 import { messageFor } from '@zestia/ember-validation';
 
 export default function present(options) {
-  return function (value) {
+  return function (value, object) {
     if (isPresent(value)) {
       return;
     }
 
-    return messageFor('present', value, options);
+    return messageFor('present', value, object, options);
   };
 }