+++ addon/-private/constraints/truthy.js
@@ -1,11 +1,11 @@
 import { messageFor } from '@zestia/ember-validation';
 
 export default function truthy(options) {
-  return function (value) {
+  return function (value, object) {
     if (value) {
       return;
     }
 
-    return messageFor('truthy', value, options);
+    return messageFor('truthy', value, object, options);
   };
 }