+++ addon/-private/messages.js
@@ -1,5 +1,6 @@
 import { get } from '@ember/object';
 import { isFunction } from '@zestia/ember-validation/-private/utils';
+const { keys } = Object;
 
 // Yep, module state
 let t = defaultMessageFn; // eslint-disable-line no-unused-vars
@@ -9,14 +10,14 @@ const locale = {
     'must-be-a-number': 'Value must be a number',
     'value-too-large': 'Value too large',
     'max-decimal-places':
-      'Value must have a maximum of {maxDecimalDigits} decimal places'
+      'Value must have a maximum of {_maxDecimalDigits} decimal places'
   },
-  date: 'Invalid date, expecting {format}',
+  date: 'Invalid date, expecting {_format}',
   email: 'Invalid email',
-  'greater-than': 'Must be greater than {otherValue}',
-  'less-than': 'Must be less than {otherValue}',
-  'max-length': 'Length is too long (max {max})',
-  'min-length': 'Length must be at least {min}',
+  'greater-than': 'Must be greater than {_value}',
+  'less-than': 'Must be less than {_value}',
+  'max-length': 'Length is too long (max {_max})',
+  'min-length': 'Length must be at least {_min}',
   number: 'Invalid number',
   'phone-number': 'Invalid phone number',
   present: 'Required value',
@@ -35,15 +36,16 @@ export function setMessageFn(fn) {
   t = fn;
 }
 
-export function messageFor(key, value, options = {}) {
+export function messageFor(key, value, object, options = {}) {
+  const tokens = keys(options).reduce(
+    (all, key) => {
+      all[`_${key}`] = options[key];
+      return all;
+    },
+    { ...object, value }
+  );
+
   const message = options.message;
-  const tokens = { ...options, value };
-
-  delete options.message;
-
-  if ('value' in options) {
-    tokens.otherValue = options.value;
-  }
 
   if (message) {
     if (isFunction(message)) {